Optimization, Bugfixing
Makes it so that the GameManager doesn't have to destroy and re-instantiate the player, but instead temporarily disables the player object, re-enabling it only when the respawn timer has passed.

Small downside that if we add a large number of variables to our prefab that can be changed within the level, we have to manually reset those settings upon re-enabling the player. It does prevent null errors in the gamemanager when trying to update using the player, though.