Pause button disables controls
Pausing the game now properly disables controls, with some special edge case code added in for if the player releases the grapple while in a paused state to prevent buggy behavior.

Makes some new prefabs based around the level spawner and game manager so they can be dragged and dropped into a scene.